Fix wrong method modifier in basic example for factory pattern
The example uses a default method in the interface, but the following
code block accesses it like it is a static method.

A default method would not make sense in this case, as you need an
instance of an implementation of the interface to call the method.
The method however is the only way to obtain such an instance and thus
can never be called.